+++ CHANGELOG.md
@@ -7,10 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Argument ``print_command_line`` to ``foxplot``
+
 ### Changed
 
 - Rename internal ``IndexedSeries`` to just ``Series``
 - Repeat previous value when there is a gap in the input
+
+### Removed
+
+- Argument ``open_new_tab`` to ``foxplot``
 
 ## [0.4.0] - 2023-10-23